host/aes_crypto.c: Add support for sdp buffer
This commit adds the host side support for a secure
buffer. Also it adds a "secure memcpy" command to
enable unencrypted data to be copied into a protected
buffer.

Some sdp_protoype code is also added for testing on
platforms where the protection isn't actually enabled
on the "secure" buffers.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>